Clean up utils into own submodules
