+++ mslib/msui/_tests/test_updater.py
@@ -29,7 +29,7 @@ import mock
 from PyQt5 import QtWidgets, QtTest
 
 from mslib.msui.updater import UpdaterUI, Updater
-from mslib.utils import Worker
+from mslib.msui.mss_qt import Worker
 
 
 def no_conda(args=None, **named_args):
@@ -91,7 +91,7 @@ class Test_MSS_ShortcutDialog:
 
     @mock.patch("subprocess.Popen", new=SubprocessDifferentVersionMock)
     @mock.patch("subprocess.run", new=SubprocessDifferentVersionMock)
-    @mock.patch("mslib.utils.Worker.create", create_mock)
+    @mock.patch("mslib.msui.mss_qt.Worker.create", create_mock)
     def test_update_recognised(self):
         self.updater.run()
 
@@ -105,7 +105,7 @@ class Test_MSS_ShortcutDialog:
 
     @mock.patch("subprocess.Popen", new=SubprocessSameMock)
     @mock.patch("subprocess.run", new=SubprocessSameMock)
-    @mock.patch("mslib.utils.Worker.create", create_mock)
+    @mock.patch("mslib.msui.mss_qt.Worker.create", create_mock)
     def test_no_update(self):
         self.updater.run()
         assert self.status == "Your MSS is up to date."
@@ -114,7 +114,7 @@ class Test_MSS_ShortcutDialog:
 
     @mock.patch("subprocess.Popen", new=SubprocessDifferentVersionMock)
     @mock.patch("subprocess.run", new=SubprocessDifferentVersionMock)
-    @mock.patch("mslib.utils.Worker.create", create_mock)
+    @mock.patch("mslib.msui.mss_qt.Worker.create", create_mock)
     def test_update_failed(self):
         self.updater.run()
         assert self.updater.new_version == "999.999.999"
@@ -126,7 +126,7 @@ class Test_MSS_ShortcutDialog:
 
     @mock.patch("subprocess.Popen", new=no_conda)
     @mock.patch("subprocess.run", new=no_conda)
-    @mock.patch("mslib.utils.Worker.create", create_mock)
+    @mock.patch("mslib.msui.mss_qt.Worker.create", create_mock)
     def test_no_conda(self):
         self.updater.run()
         assert self.updater.new_version is None and self.updater.old_version is None
@@ -135,7 +135,7 @@ class Test_MSS_ShortcutDialog:
 
     @mock.patch("subprocess.Popen", new=no_conda)
     @mock.patch("subprocess.run", new=no_conda)
-    @mock.patch("mslib.utils.Worker.create", create_mock)
+    @mock.patch("mslib.msui.mss_qt.Worker.create", create_mock)
     def test_exception(self):
         self.updater.new_version = "999.999.999"
         self.updater.old_version = "999.999.999"
@@ -146,7 +146,7 @@ class Test_MSS_ShortcutDialog:
     @mock.patch("subprocess.Popen", new=SubprocessSameMock)
     @mock.patch("subprocess.run", new=SubprocessSameMock)
     @mock.patch("PyQt5.QtWidgets.QMessageBox.information", return_value=QtWidgets.QMessageBox.Yes)
-    @mock.patch("mslib.utils.Worker.create", create_mock)
+    @mock.patch("mslib.msui.mss_qt.Worker.create", create_mock)
     def test_ui(self, mock):
         ui = UpdaterUI()
         ui.updater.on_update_available.emit("", "")
