+++ mslib/msui/airdata_dockwidget.py
@@ -27,7 +27,8 @@
 import pycountry
 from mslib.msui.mss_qt import ui_airdata_dockwidget as ui
 from PyQt5 import QtWidgets, QtCore
-from mslib.utils import save_settings_qsettings, load_settings_qsettings, _airspace_cache, update_airspace, get_airports
+from mslib.utils.config import save_settings_qsettings, load_settings_qsettings
+from mslib.utils.airdata import _airspace_cache, update_airspace, get_airports
 
 
 class AirdataDockwidget(QtWidgets.QWidget, ui.Ui_AirdataDockwidget):
