+++ mslib/msui/flighttrack.py
@@ -44,9 +44,11 @@ import xml.parsers.expat
 
 from PyQt5 import QtGui, QtCore, QtWidgets
 
-from mslib import utils, __version__
+from mslib import __version__
 from mslib.utils.units import units
-from mslib.utils import thermolib, config_loader, find_location, save_settings_qsettings, load_settings_qsettings
+from mslib.utils import thermolib
+from mslib.utils.coordinate import find_location, path_points, get_distance
+from mslib.utils.config import config_loader, save_settings_qsettings, load_settings_qsettings
 from mslib.msui import MissionSupportSystemDefaultConfig as mss_default
 from mslib.msui.mss_qt import variant_to_string, variant_to_float
 from mslib.msui.performance_settings import DEFAULT_PERFORMANCE
@@ -257,7 +259,7 @@ class WaypointsTableModel(QtCore.QAbstractTableModel):
         Returns lats, lons.
         """
         path = [[wp.lat, wp.lon, wp.utc_time] for wp in self.waypoints]
-        return utils.path_points(
+        return path_points(
             path, numpoints=numpoints, connection=connection)
 
     def headerData(self, section, orientation, role=QtCore.Qt.DisplayRole):
@@ -489,8 +491,8 @@ class WaypointsTableModel(QtCore.QAbstractTableModel):
                 wp1.ascent_rate = 0
             else:
                 wp0 = waypoints[pos - 1]
-                wp1.distance_to_prev = utils.get_distance((wp0.lat, wp0.lon),
-                                                          (wp1.lat, wp1.lon))
+                wp1.distance_to_prev = get_distance((wp0.lat, wp0.lon),
+                                                    (wp1.lat, wp1.lon))
 
                 last = (pos - 1 == rows)
                 time, fuel = get_duration_fuel(
@@ -510,8 +512,8 @@ class WaypointsTableModel(QtCore.QAbstractTableModel):
         # Update the distance of the following waypoint as well.
         if pos < len(waypoints) - 1:
             wp2 = waypoints[pos + 1]
-            wp2.distance_to_prev = utils.get_distance((wp1.lat, wp1.lon),
-                                                      (wp2.lat, wp2.lon))
+            wp2.distance_to_prev = get_distance((wp1.lat, wp1.lon),
+                                                (wp2.lat, wp2.lon))
             if wp2.leg_time != 0:
                 wp2.ascent_rate = int((wp2.flightlevel - wp1.flightlevel) * 100 / (wp2.leg_time / 60))
             else:
