+++ mslib/msui/kmloverlay_dockwidget.py
@@ -35,7 +35,7 @@ from matplotlib import patheffects
 from mslib.msui.mss_qt import get_open_filenames, get_save_filename
 from mslib.msui.mss_qt import ui_kmloverlay_dockwidget as ui
 from PyQt5 import QtGui, QtWidgets, QtCore
-from mslib.utils import save_settings_qsettings, load_settings_qsettings
+from mslib.utils.config import save_settings_qsettings, load_settings_qsettings
 
 
 class KMLPatch(object):
