+++ mslib/msui/mpl_pathinteractor.py
@@ -52,7 +52,7 @@ import matplotlib.path as mpath
 import matplotlib.patches as mpatches
 from PyQt5 import QtCore, QtWidgets
 
-from mslib.utils import get_distance, find_location, path_points, latlon_points
+from mslib.utils.coordinate import get_distance, find_location, path_points, latlon_points
 from mslib.utils.units import units
 from mslib.utils.thermolib import pressure2flightlevel
 from mslib.msui import flighttrack as ft
