+++ mslib/msui/mscolab_admin_window.py
@@ -31,7 +31,8 @@ from werkzeug.urls import url_join
 
 from PyQt5 import QtCore, QtWidgets
 from mslib.msui.qt5 import ui_mscolab_admin_window as ui
-from mslib.utils import config_loader, show_popup
+from mslib.msui.mss_qt import show_popup
+from mslib.utils.config import config_loader
 
 
 class MSColabAdminWindow(QtWidgets.QMainWindow, ui.Ui_MscolabAdminWindow):
