+++ mslib/msui/mss_pyui.py
@@ -58,10 +58,11 @@ from mslib.msui import constants
 from mslib.msui import wms_control
 from mslib.msui import mscolab
 from mslib.msui.updater import UpdaterUI
-from mslib.utils import config_loader, setup_logging, Worker, Updater
+from mslib.utils import setup_logging
+from mslib.utils.config import config_loader
 from mslib.plugins.io.csv import load_from_csv, save_to_csv
 from mslib.msui.icons import icons, python_powered
-from mslib.msui.mss_qt import get_open_filename, get_save_filename
+from mslib.msui.mss_qt import get_open_filename, get_save_filename, Worker, Updater
 from PyQt5 import QtGui, QtCore, QtWidgets
 
 # Add config path to PYTHONPATH so plugins located there may be found
