+++ mslib/msui/remotesensing_dockwidget.py
@@ -37,7 +37,8 @@ import skyfield_data
 from mslib.msui.constants import MSS_CONFIG_PATH
 from PyQt5 import QtGui, QtWidgets
 from mslib.msui.mss_qt import ui_remotesensing_dockwidget as ui
-from mslib.utils import jsec_to_datetime, datetime_to_jsec, get_distance, rotate_point, fix_angle
+from mslib.utils.time import jsec_to_datetime, datetime_to_jsec
+from mslib.utils.coordinate import get_distance, rotate_point, fix_angle
 
 
 EARTH_RADIUS = 6371.
