+++ mslib/msui/tableview.py
@@ -38,10 +38,10 @@ from mslib.msui import hexagon_dockwidget as hex
 from mslib.msui import performance_settings as perfset
 from PyQt5 import QtWidgets, QtGui
 from mslib.msui.mss_qt import ui_tableview_window as ui
+from mslib.msui.mss_qt import dropEvent, dragEnterEvent
 from mslib.msui import flighttrack as ft
 from mslib.msui.viewwindows import MSSViewWindow
 from mslib.msui.icons import icons
-from mslib.utils import dropEvent, dragEnterEvent
 
 try:
     import mpl_toolkits.basemap.pyproj as pyproj
