+++ mslib/msui/topview.py
@@ -30,7 +30,8 @@
 
 import functools
 import logging
-from mslib.utils import config_loader, get_projection_params, save_settings_qsettings, load_settings_qsettings
+from mslib.utils.config import config_loader, save_settings_qsettings, load_settings_qsettings
+from mslib.utils.coordinate import get_projection_params
 from PyQt5 import QtGui, QtWidgets, QtCore
 from mslib.msui.mss_qt import ui_topview_window as ui
 from mslib.msui.mss_qt import ui_topview_mapappearance as ui_ma
